+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - VFX: Fixed LPPV with lit particles in deferred (case 1293608)
 - Fixed computation of geometric normal in path tracing (case 1293029).
 - Fixed issues with path-traced volumetric scattering (cases 1295222, 1295234).
+- Fixed an issue with the frame count management for the volumetric fog (case 1299251).
 
 ### Changed
 - Removed the material pass probe volumes evaluation mode.